% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_units}
\alias{add_units}
\title{Adds units from labels}
\usage{
add_units(df)
}
\arguments{
\item{df}{}
}
\description{
\verb{labelr::labels$set needs to be called first} such that the function can
retrieve the values. Adding the units is most likely the last step!
}
