% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_candidate_set.R
\name{generate_candidate_set}
\alias{generate_candidate_set}
\title{Generates a candidate set for implicit partial profile}
\usage{
generate_candidate_set(keep = NULL, names = TRUE, ngene = FALSE)
}
\arguments{
\item{keep}{random sample to keep (defines nrows of output)}

\item{names}{should "A_" and "B_" prefix be added for bundles (defaults to \code{TRUE}).}

\item{ngene}{The ngene user manual suggests, that the candidature set does not
have these prefixes, but repeated col names for the bundles (defaults to \code{FALSE}).}
}
\value{
ngene candidate set (conforms to examples in ngene user manual)
}
\description{
Generates a candidate set for implicit partial profile
}
\seealso{
ngene user manual
}
