% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_import_module.R
\name{python_import_module}
\alias{python_import_module}
\title{Load python modules from \code{./python} repo}
\usage{
python_import_module(module)
}
\arguments{
\item{module}{module name without \code{.py} ending}
}
\description{
Load python modules from \code{./python} repo
}
\examples{
trump <- python_import_module("trump")
trump$make("America")
trump$great("again")
}
