% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_effect_codes.R
\name{replace_codes_from_archetypes}
\alias{replace_codes_from_archetypes}
\title{Replaces effect codes with values from archetypes.}
\usage{
replace_codes_from_archetypes(df, archetypes)
}
\arguments{
\item{df}{labeled design}

\item{archetypes}{see data-raw (generated with help of \code{tcsscraper})}
}
\description{
Should be called after \code{labelr::label_df} has been applied
}
\seealso{
\code{tcsscraper} python package (in particular jupyter notebook)
}
