% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_effect_codes.R
\name{replace_effect_codes}
\alias{replace_effect_codes}
\title{Replace effect codes with human readable levels}
\usage{
replace_effect_codes(design, add_units = TRUE)
}
\arguments{
\item{design}{output from \code{experiments::generate_design()$final}}

\item{add_units}{units should be added for choice cards}
}
\description{
Master function that calls all the helpers
}
