% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_effect_codes.R
\name{sub_codes_from_archs}
\alias{sub_codes_from_archs}
\title{Replaces effect codes with values from archetypes.}
\usage{
sub_codes_from_archs(df, archetypes)
}
\arguments{
\item{df}{labeled design}

\item{archetypes}{see \code{generate_generic_archetypes} and \code{gen_archs}}
}
\description{
Should be called after \code{labelr::label_df} has been applied
}
\seealso{
\code{tcsscraper} python package (in particular sub-package \code{experiments})
}
